- struct の field / positional param / __copy と IlExport の field・method 名を
  写像後 (snake_case) の名前にし、hot reload の migration と on_reload 呼び出し、
  tcs2c の entry 探索を同じ名前に揃える
- TestHelper の luaExpr は C# の名前で書けるようにし、写像は helper が行う
  (master 側で増えたテストをそのまま通す)。raw script のテストと fuzz reload の
  harness は写像後の名前に書き換え
- IlBuild.Expressions を 800 行以下に分割 (補間文字列を別 partial へ)
- T231 / T232 が master と重複していたので T238 / T239 に採番し直す

+++ Transpiler.Tests/LuaNamingTests.cs
@@ -153,7 +153,7 @@ public class LuaNamingTests
         public class PassOpts
         {
             public TextureRef? Target;
-            public double[]? ClearColor;
+            public float[]? ClearColor;
             public int? Load;
         }
         public class Readback
@@ -192,7 +192,7 @@ public class LuaNamingTests
                     Gfx.BeginPass(new PassOpts
                     {
                         Target = Gfx.MainTex,
-                        ClearColor = new double[] { 0.5, 1.0 },
+                        ClearColor = new float[] { 0.5f, 1.0f },
                         Load = (int)Gfx.LoadAction.DontCare,
                     });
                     var rb = Gfx.Readback();
